feat: 1450 account module refactoring
