+++ packages/ethers-adapter/tests/adapters/contract-adapter.solo.test.ts
@@ -29,7 +29,7 @@ describe('Hardhat contract adapter tests', () => {
      * Init thor client and provider before each test
      */
     beforeEach(() => {
-        thorClient = ThorClient.fromUrl(THOR_SOLO_URL);
+        thorClient = ThorClient.at(THOR_SOLO_URL);
         provider = new HardhatVeChainProvider(
             new ProviderInternalBaseWallet([]),
             THOR_SOLO_URL,
