+++ packages/ethers-adapter/tests/adapters/factory-adapter.solo.test.ts
@@ -27,7 +27,7 @@ describe('Hardhat factory adapter tests', () => {
      * Init thor client and provider before each test
      */
     beforeEach(() => {
-        thorClient = ThorClient.fromUrl(THOR_SOLO_URL);
+        thorClient = ThorClient.at(THOR_SOLO_URL);
         provider = new HardhatVeChainProvider(
             new ProviderInternalBaseWallet([]),
             THOR_SOLO_URL,
