+++ packages/network/src/thor-client/accounts/index.ts
@@ -1,2 +1,2 @@
 export type * from './types.d';
-export * from './accounts-module';
+export * from './AccountsModule';
